+++ tools/ledger/src/main/resources/commands
@@ -21,4 +21,4 @@ org.apache.bookkeeper.tools.cli.commands.BookieIdCommandGroup
 org.apache.bookkeeper.tools.cli.commands.BookiesCommandGroup
 org.apache.bookkeeper.tools.cli.commands.CookieCommandGroup
 org.apache.bookkeeper.tools.cli.commands.LedgerCommandGroup
-org.apache.bookkeeper.tools.cli.commands.ToggleCommandGroup
+org.apache.bookkeeper.tools.cli.commands.AutoRecoveryCommandGroup
